# [Performance] Batch Send from Tokenizer Manager.